[Order] Some crucial and PR review fixes
