+++ src/Sylius/Bundle/CoreBundle/Resources/config/doctrine/model/OrderItemUnit.orm.xml
@@ -28,10 +28,6 @@
             <gedmo:timestampable on="update"/>
         </field>
 
-        <many-to-one field="orderItem" target-entity="Sylius\Component\Core\Model\OrderItem" inversed-by="itemUnits">
-            <join-column name="order_item_id" referenced-column-name="id" nullable="false" on-delete="CASCADE" />
-        </many-to-one>
-
         <many-to-one field="shipment" target-entity="Sylius\Component\Shipping\Model\ShipmentInterface" inversed-by="items">
             <join-column name="shipment_id" referenced-column-name="id" nullable="true" on-delete="SET NULL" />
         </many-to-one>
