+++ src/Sylius/Bundle/InventoryBundle/Resources/config/doctrine/model/InventoryUnit.orm.xml
@@ -24,10 +24,6 @@
 
         <field name="inventoryState" column="inventory_state" type="string" />
 
-        <many-to-one field="stockable" target-entity="Sylius\Component\Core\Model\ProductVariant">
-            <join-column name="stockable_id" referenced-column-name="id" nullable="false" />
-        </many-to-one>
-
         <field name="createdAt" column="created_at" type="datetime">
             <gedmo:timestampable on="create"/>
         </field>
