+++ src/Sylius/Bundle/OrderBundle/Resources/config/doctrine/model/OrderItem.orm.xml
@@ -42,6 +42,12 @@
             </cascade>
         </one-to-many>
 
+        <one-to-many field="itemUnits" target-entity="Sylius\Component\Order\Model\OrderItemUnit" mapped-by="orderItem" orphan-removal="true">
+            <cascade>
+                <cascade-all/>
+            </cascade>
+        </one-to-many>
+
         <gedmo:loggable />
     </mapped-superclass>
