+++ src/Sylius/Bundle/OrderBundle/Resources/config/doctrine/model/OrderItemUnit.orm.xml
@@ -24,6 +24,10 @@
         <field name="total" column="total" type="integer" />
         <field name="adjustmentsTotal" column="adjustments_total" type="integer" />
 
+        <many-to-one field="orderItem" target-entity="Sylius\Component\Order\Model\OrderItem" inversed-by="itemUnits">
+            <join-column name="order_item_id" referenced-column-name="id" nullable="false" on-delete="CASCADE" />
+        </many-to-one>
+
         <one-to-many field="adjustments" target-entity="Sylius\Component\Order\Model\AdjustmentInterface" mapped-by="orderItemUnit" orphan-removal="true">
             <cascade>
                 <cascade-all/>
