+++ src/Sylius/Component/Inventory/Factory/InventoryUnitFactory.php
@@ -36,7 +36,6 @@ class InventoryUnitFactory extends Factory implements InventoryUnitFactoryInterf
 
         for ($i = 0; $i < $quantity; $i++) {
             $inventoryUnit = $this->createNew();
-            $inventoryUnit->setStockable($stockable);
             $inventoryUnit->setInventoryState($state);
 
             $units->add($inventoryUnit);
